# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Content Atlas API",
    "description": "A data consolidation platform for SMBs to consolidate data from multiple sources into PostgreSQL",
    "version": "1.0.1"
  },
  "paths": {
...
    "/tables/{table_name}": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Query Table",
        "description": "Query table data with pagination, optional search, sorting, and filtering.\n\nParameters:\n- import_id: Optional UUID to filter rows by specific import",
        "operationId": "query_table_tables__table_name__get",
        "parameters": [
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "import_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Import Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Sort Order"
            }
          },
          {
            "name": "filters",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON list of filters: [{\"column\":\"status\",\"operator\":\"eq\",\"value\":\"Active\"}]",
              "title": "Filters"
            },
            "description": "JSON list of filters: [{\"column\":\"status\",\"operator\":\"eq\",\"value\":\"Active\"}]"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tables"
        ],
        "summary": "Delete Table",
        "description": "Delete a user-created table and clean up all related records.\n\nThis endpoint:\n1. Drops the table from the database\n2. Cleans up import_history records\n3. Cleans up file_imports records\n4. Resets uploaded_files status to 'uploaded' (unmapped)\n\nParameters:\n- table_name: Name of the table to delete\n\nReturns:\n- Success status and cleanup summary",
        "operationId": "delete_table_tables__table_name__delete",
        "parameters": [
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_name}/export": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Export Table",
        "description": "Stream the full contents of a user table as CSV.\n\nThe export excludes metadata columns (prefixed with \"_\") and returns a streaming\nCSV response so large tables do not exhaust memory.",
        "operationId": "export_table_tables__table_name__export_get",
        "parameters": [
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1000000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional maximum number of rows to include in the export",
              "title": "Limit"
            },
            "description": "Optional maximum number of rows to include in the export"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_name}/schema": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Get Table Schema",
        "description": "Get table column information.\n\nReturns detailed schema information for a specific table, including\ncolumn names, data types, and nullability.\n\nParameters:\n- table_name: Name of the table\n\nReturns:\n- Table schema with column details",
        "operationId": "get_table_schema_tables__table_name__schema_get",
        "parameters": [
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_name}/stats": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "Get Table Stats",
        "description": "Get basic table statistics.\n\nReturns summary statistics for a table including row count,\ncolumn count, and data type distribution.\n\nParameters:\n- table_name: Name of the table\n\nReturns:\n- Table statistics",
        "operationId": "get_table_stats_tables__table_name__stats_get",
        "parameters": [
          {
            "name": "table_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tables": {
      "get": {
        "tags": [
          "public-api"
        ],
        "summary": "List Available Tables",
        "description": "Lists all user-accessible tables along with their current row counts.",
        "operationId": "public_list_tables_endpoint_api_v1_tables_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablesListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TableDataResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "success",
          "table_name",
          "data",
          "total_rows",
          "limit",
          "offset"
        ],
        "title": "TableDataResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}